requirements.txt
absl-py==2.1.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.6.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attr==0.3.1
attrs==25.1.0
azure-core==1.32.0
azure-storage-blob==12.24.1
babel==2.17.0
bayesian-torch==0.5.0
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.37.8
botocore==1.37.8
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
contourpy==1.3.1
crc32c==2.7.1
cryptography==44.0.2
cycler==0.12.1
dask==2025.2.0
datamodel-code-generator==0.26.1
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.7
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
docutils==0.21.2
donfig==0.8.1.post1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
expiringdict==1.2.2
Faker==36.2.2
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
genson==1.3.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.1
google-cloud-core==2.4.2
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
imagecodecs==2024.12.30
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.5.0
inflect==5.6.2
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
kymatio==0.3.0
label-studio==1.16.0
label-studio-sdk==1.0.10
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
lockfile==0.12.2
lxml==5.3.1
lxml_html_clean==0.4.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numcodecs==0.15.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.65.4
opencv-python==4.11.0.86
openpyxl==3.1.5
openslide-bin==4.0.0.6
openslide-python==1.4.1
opentelemetry-api==1.30.0
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.0.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pdoc3==0.11.6
pexpect==4.9.0
pillow==11.1.0
pkgconfig==1.5.5
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyRFC3339==2.0.1
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
pyvips==2.2.3
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.23.1
rq==1.16.2
rsa==4.9
rstr==3.2.2
ruff==0.9.9
rules==3.4
s3transfer==0.11.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.22.0
setuptools==75.8.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tifftools==1.5.2
tldextract==5.1.3
toml==0.10.2
toolz==1.0.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Tree==0.2.4
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
-e git+https://github.com/mattiabestiaccia/Wavelet-Scattering-Trasform-Analysis.git@38799398dac70b75534648d6634a8845949c5667#egg=wavelet_lib
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.40.0
wrapt==1.17.2
xmljson==0.2.1
zarr==3.0.4
zipp==3.21.0
